requirements.txt
# Explainable Multimodal Clinical AI Assistant
# Compatible with Apple Silicon (M-series) — no CUDA required

# --- Core ML ---
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.26.0
Pillow>=10.0.0
scikit-learn>=1.3.0

# --- FastAPI Backend ---
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# --- Streamlit Frontend ---
streamlit>=1.30.0

# --- MedGemma: Local MLX Inference ---
mlx>=0.25.0
mlx-vlm>=0.2.0

# --- MedGemma: API Fallback ---
google-generativeai>=0.5.0

# --- Explainability & Visualization ---
matplotlib>=3.8.0
opencv-python-headless>=4.9.0

# --- Utilities ---
python-dotenv>=1.0.0
httpx>=0.27.0
